Added simple splitter component, with vertical and horizontal demos.
Note: The current implementation requires the user to ensure the correct markup is used, to make this more flexible would require changes to the splitter control itself (or maybe my understanding of it).

However, this works as a starting point.

Also added some Visual Studio files, hope that is okay.